+++ Project3/src/main/java/com/Project3/Project3/repository/ReviewRepository.java
@@ -7,10 +7,9 @@ import org.springframework.stereotype.Repository;
 import com.Project3.Project3.model.Review;
 
 @Repository
-public interface ReviewRepository extends CrudRepository<Review, Integer> {
+public interface ReviewRepository extends CrudRepository<Review, Long> {
 
     // Corrected query to access TravelPackage ID.  Use JPQL.
     @Query("SELECT r FROM Review r WHERE r.package1.packageId = :packageId")
     List<Review> findByPackage1_PackageId(int packageId);
-
 }
